backend-calmera/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.0
pydantic==2.7.0
pydantic[email]==2.7.0
python-multipart==0.0.9
redis==5.0.1
phonenumbers==8.13.30
python-dotenv==1.0.1
pytest==8.0.1
httpx==0.26.0
transformers==4.37.2
requests-toolbelt
pandas==2.2.0
openai==1.7.1 